refactor(core): migrate to new files API
